requirements.txt
streamlit
tensorflow
opencv-python
umpy
pandas
matplotlib
fpdf
